pyproject.toml
[project]
name = "TestSmith"
version = "0.1.0"
description = "TestSmith is an intelligent agent that automatically transforms software requirements into a comprehensive suite of test cases, ensuring robust quality and accelerating product delivery."
readme = "README.md"
keywords = ["QA", "Testing", "Coverage", "Unit", "LangChain"]
requires-python = ">=3.11"
dependencies = [
    # Core dependencies for LLM functionality
    "anthropic>=0.49.0",
    "langchain>=0.3.20",
    "langchain-core>=0.3.41",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.7",
    "langchain-anthropic>=0.3.9",
    "langgraph>=0.3.5",
    "openai>=1.65.4",
    # Testing and validation dependencies
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    # Utilities
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "structlog>=25.1.0",
    "gitpython>=3.1.44",
    "build>=1.2.2.post1",
    "pytest-mock>=3.14.0",
]

[project.optional-dependencies]
dev = [
    # Development tools
    "mypy>=1.15.0",
    "types-psutil>=7.0.0.20250218",
    "black>=25.1.0",
    "isort>=6.0.1",
    "flake8>=7.1.2",
    "pytest-mock>=3.14.0",
]

[project.scripts]
qa-agent = "qa_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["qa_agent"]

[[project.authors]]
name = "QA Agent Team"
email = "qaagent@example.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.cache
  | generated_tests
  | test_output
  | temp
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = [".git/*", "__pycache__/*", "*.pyi", "generated_tests/*", "test_output/*", "temp/*"]
known_first_party = ["qa_agent"]

[tool.semantic_release]
version_variable = "pyproject.toml:version"
branch = "main"
upload_to_pypi = false
upload_to_release = true
build_command = "pip install build && python -m build"

